refactor: improve form queries

+++ backend/src/main/java/format/backend/form/dto/FormDetailResponseDto.java
@@ -46,7 +46,7 @@ public record FormDetailResponseDto(
 
         @Nullable Double ratingAvg,
 
-        @Nullable Double userRating,
+        @Nullable Integer userRating,
 
         @Nullable String authorName,
